GoPro.Plus.Auth

Description

GoPro Plus authentication.

Documentation

authenticate

Arguments

:: MonadIO m

=> String
Email/username

-> String
Password

-> m AuthInfo

Authenticate against the GoPro Plus service.

refreshAuth :: MonadIO m => AuthInfo -> m AuthInfo

Refresh authentication credentials using a refresh token.

data AuthInfo

An Authentication response.

Constructors

AuthInfo

Fields

_access_token :: Token

_expires_in :: Int

_refresh_token :: Text

_resource_owner_id :: Text

access_token :: Lens' AuthInfo Token

expires_in :: Lens' AuthInfo Int

refresh_token :: Lens' AuthInfo Text

resource_owner_id :: Lens' AuthInfo Text

class Monad m => HasGoProAuth m where

A Monad may have a HasGoProAuth instance to indicate it knows how to authenticate against the GoPro Plus service.

Methods

goproAuth :: m AuthInfo

Get the GoPro AuthInfo to use.

withAuth :: AuthInfo -> AuthReader m a -> m a

Convenient function for passing around auth info. You probably don't want to uset his, but it can be convenient when experimenting.

type AuthReader = ReaderT AuthInfo

AuthReader is a convenience type that's useful for doing small experiments where you don't already have your own Reader or similar. e.g., in ghci you might type:

(m :: Medium) <- withAuth (AuthInfo accessToken 0 "" "") $ medium mediumID

type Token = Text