GoPro.Plus.Media

Description

GoPro Plus media client.

Accessing Data

list

Arguments

:: (HasGoProAuth m, MonadIO m)

=> Int
Number of items per page.

-> Int
Page number (one-based).

-> m ([Medium], PageInfo)

List a page worth of media.

listAll :: (HasGoProAuth m, MonadIO m) => m [Medium]

List all media.

listWhile :: (HasGoProAuth m, MonadIO m) => ([Medium] -> Bool) -> m [Medium]

List all media while returned batches pass the given predicate.

medium :: (HasGoProAuth m, FromJSON j, MonadIO m) => MediumID -> m j

Get the current Medium record for the given Medium ID.

notReady :: (HasGoProAuth m, MonadIO m) => m [Medium]

Get a list of items whose processing is failed or incomplete.

This includes items that are currently uploading, or items that will not upload.

retrieve :: (HasGoProAuth m, FromJSON j, MonadIO m) => MediumID -> m j

Get download descriptors for a given medium. The format is typically FileInfo, but it can be useful to map it into something else.

delete :: (HasGoProAuth m, MonadIO m) => MediumID -> m [Error]

Delete an item.

reprocess :: (HasGoProAuth m, MonadIO m) => MediumID -> m ()

Reprocess a failed upload for the given medium ID

fetchThumbnail :: (HasGoProAuth m, MonadIO m) => Medium -> m ByteString

Fetch a 450px wide thumbnail data for the given medium.

Data Types

data PageInfo

Pagination info returned from lists.

Constructors

PageInfo

Fields

_current_page :: Int

_per_page :: Int

_total_items :: Int

_total_pages :: Int

current_page :: Lens' PageInfo Int

per_page :: Lens' PageInfo Int

total_items :: Lens' PageInfo Int

total_pages :: Lens' PageInfo Int

type MediumID = Text

GoPro-assigned identifier for an uploaded item.

data MediumType

Type of Media for a given item.

Constructors

Photo
a still photo

Video
normal video

TimeLapse
a timelapse series of photos

TimeLapseVideo
a timelapse video

Burst
a set of photos taken in a burst

Chaptered

Livestream

Looped

LoopedVideo

BurstVideo

Continuous

ExternalVideo

Session

MultiClipEdit

data ReadyToViewType

Constructors

ViewReady

ViewFailure

ViewLoading

ViewRegistered

ViewTranscoding

ViewProcessing

ViewUploading

ViewPreTranscoding

ViewUpdating

data Medium

Constructors

Medium

Fields

_medium_id :: MediumID

_medium_camera_model :: Maybe String

_medium_captured_at :: UTCTime

_medium_created_at :: UTCTime

_medium_file_size :: Maybe Int

_medium_moments_count :: Int

_medium_ready_to_view :: ReadyToViewType

_medium_source_duration :: Maybe String

_medium_type :: MediumType

_medium_token :: String

_medium_width :: Maybe Int

_medium_height :: Maybe Int

_medium_filename :: Maybe String

medium_id :: Lens' Medium MediumID

medium_camera_model :: Lens' Medium (Maybe String)

medium_captured_at :: Lens' Medium UTCTime

medium_created_at :: Lens' Medium UTCTime

medium_file_size :: Lens' Medium (Maybe Int)

medium_moments_count :: Lens' Medium Int

medium_ready_to_view :: Lens' Medium ReadyToViewType

medium_source_duration :: Lens' Medium (Maybe String)

medium_type :: Lens' Medium MediumType

medium_token :: Lens' Medium String

medium_width :: Lens' Medium (Maybe Int)

medium_height :: Lens' Medium (Maybe Int)

medium_filename :: Lens' Medium (Maybe String)

data Listing

Constructors

Listing

Fields

_media :: [Medium]

_pages :: PageInfo

media :: Lens' Listing [Medium]

pages :: Lens' Listing PageInfo

class HasMediaURL c where

Methods

media_url :: Lens' c String

class HasMediaHead c where

Methods

media_head :: Lens' c String

class HasMediaLabel c where

Methods

media_label :: Lens' c String

class HasMediaType c where

Methods

media_type :: Lens' c String

class HasMediaItemNumber c where

Methods

media_item_number :: Lens' c (Maybe Int)

data File

Constructors

File

Fields

_file_camera_position :: String

_file_height :: Int

_file_width :: Int

_file_item_number :: Int

_file_orientation :: Int

_file_transforms :: Maybe [String]

_file_head :: String

_file_url :: String

file_camera_position :: Lens' File String

file_height :: Lens' File Int

file_width :: Lens' File Int

file_item_number :: Lens' File Int

file_orientation :: Lens' File Int

file_head :: Lens' File String

file_url :: Lens' File String

file_transforms :: Lens' File (Maybe [String])

data Variation

Constructors

Variation

Fields

_var_height :: Int

_var_width :: Int

_var_label :: String

_var_quality :: String

_var_transforms :: Maybe [String]

_var_item_number :: Maybe Int

_var_type :: String

_var_head :: String

_var_url :: String

var_height :: Lens' Variation Int

var_width :: Lens' Variation Int

var_label :: Lens' Variation String

var_quality :: Lens' Variation String

var_type :: Lens' Variation String

var_transforms :: Lens' Variation (Maybe [String])

var_head :: Lens' Variation String

var_url :: Lens' Variation String

var_item_number :: Lens' Variation (Maybe Int)

data SpriteFrame

Constructors

SpriteFrame

Fields

_frame_count :: Int

_frame_height :: Int

_frame_width :: Int

frame_count :: Lens' SpriteFrame Int

frame_height :: Lens' SpriteFrame Int

frame_width :: Lens' SpriteFrame Int

data Sprite

Constructors

Sprite

Fields

_sprite_fps :: Double

_sprite_frame :: SpriteFrame

_sprite_height :: Int

_sprite_width :: Int

_sprite_type :: String

_sprite_heads :: [String]

_sprite_urls :: [String]

sprite_fps :: Lens' Sprite Double

sprite_frame :: Lens' Sprite SpriteFrame

sprite_height :: Lens' Sprite Int

sprite_width :: Lens' Sprite Int

sprite_type :: Lens' Sprite String

sprite_heads :: Lens' Sprite [String]

sprite_urls :: Lens' Sprite [String]

data SidecarFile

Constructors

SidecarFile

Fields

_sidecar_fps :: Double

_sidecar_label :: String

_sidecar_type :: String

_sidecar_head :: String

_sidecar_url :: String

_sidecar_item_number :: Maybe Int

sidecar_fps :: Lens' SidecarFile Double

sidecar_label :: Lens' SidecarFile String

sidecar_type :: Lens' SidecarFile String

sidecar_head :: Lens' SidecarFile String

sidecar_url :: Lens' SidecarFile String

sidecar_item_number :: Lens' SidecarFile (Maybe Int)

data FileStuff

Constructors

FileStuff

Fields

_files :: [File]

_variations :: [Variation]

_sprites :: [Sprite]

_sidecar_files :: [SidecarFile]

files :: Lens' FileStuff [File]

variations :: Lens' FileStuff [Variation]

sprites :: Lens' FileStuff [Sprite]

sidecar_files :: Lens' FileStuff [SidecarFile]

data FileInfo

Constructors

FileInfo

Fields

_fileStuff :: FileStuff

_filename :: String

fileStuff :: Lens' FileInfo FileStuff

filename :: Lens' FileInfo String

data Error

Constructors

Error

Fields

_error_reason :: String

_error_code :: Int

_error_description :: String

_error_id :: String

error_reason :: Lens' Error String

error_code :: Lens' Error Int

error_description :: Lens' Error String

error_id :: Lens' Error String

data Moment

A moment of interestingness in a Medium.

Constructors

Moment

Fields

_moment_id :: Text

_moment_time :: Maybe Int

moment_id :: Lens' Moment Text

moment_time :: Lens' Moment (Maybe Int)

moments :: (HasGoProAuth m, MonadIO m) => MediumID -> m [Moment]

Get the moments for the given medium.

Low-level Junk

updateMedium

Arguments

:: (HasGoProAuth m, MonadIO m, FromJSON j, Putable a)

=> (j -> a)
Transformation function.

-> MediumID
Medium to update.

-> m ()

Fetch, modify, and store a medium value.

putMedium :: (HasGoProAuth m, MonadIO m, Putable a) => MediumID -> a -> m ()

Put a Medium. It's probably best to get a raw JSON Value and update it in place.